[sailesh] Add: T-Sql and MultiThreading implementation

+++ EmployeePayrollService/EmployeeModel.cs
@@ -14,12 +14,12 @@ namespace EmployeePayrollService
         public string Address { get; set; }
         public string Department { get; set; }
         public char Gender { get; set; }
-        public decimal BasicPay { get; set; }
-        public decimal Salary { get; set; }
-        public decimal Deductions { get; set; }
+        public double BasicPay { get; set; }
+        public double Salary { get; set; }
+        public double Deductions { get; set; }
         public double TaxablePay { get; set; }
         public double Tax { get; set; }
-        public decimal NetPay { get; set; }
+        public double NetPay { get; set; }
         public DateTime StartDate { get; set; }
